projects/gui-proj/pyproject.toml
[project]
name = "gui-proj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "geopy>=2.4.1",
    "kafka-python-ng>=2.2.3",
    "nicegui>=2.13.0",
    "plotly>=6.0.1",
]
